+++ packages/common/src/index.ts
@@ -8,5 +8,6 @@ dotenv.config({ path: path.resolve("../../.env") });
 import { consts } from "./consts";
 import { globalSchema, validateEnvVars } from "./env";
 import { tsupConfig } from "./tsup.config";
+import * as api from "./rest";
 
-export { consts, validateEnvVars, globalSchema, tsupConfig };
+export { consts, validateEnvVars, globalSchema, tsupConfig, api };
