+++ packages/common/src/rest/channels/index.ts
@@ -1,0 +1,1 @@
+export * as channels from "./get";
