+++ packages/common/src/rest/test.ts
@@ -1,4 +1,4 @@
-import api from "./index";
+import api from "./request";
 
 try {
     console.log(await api.guilds.auditLogs.get({ "guild.id": "849562494447124500", queryParams: { action_type: 24 } }));
