+++ services/gateway/src/env.ts
@@ -1,0 +1,5 @@
+import { validateEnvVars } from "@yin/common";
+
+const env = validateEnvVars();
+
+export { env };
