+++ services/gateway/src/events/default.ts
@@ -1,0 +1,5 @@
+import { DiscordPacket } from "../packets/BasePacket";
+
+export default (_packet: DiscordPacket<any>) => {
+    console.log("No handler found for this event.");
+};
