+++ services/gateway/src/index.ts
@@ -1,10 +1,9 @@
 import sourceMapSupport from "source-map-support";
 sourceMapSupport.install();
 
-import { WebSocket } from "./WebSocket";
-import { validateEnvVars } from "@yin/common";
+import("./env");
 
-validateEnvVars();
+import { WebSocket } from "./WebSocket";
 
 const socket = new WebSocket();
 socket.connect();
